+++ ee/app/license/definitions/ILicense.ts
@@ -1,0 +1,11 @@
+import { ILicenseTag } from './ILicenseTag';
+
+export interface ILicense {
+	url: string;
+	expiry: string;
+	maxActiveUsers: number;
+	modules: string[];
+	maxGuestUsers: number;
+	maxRoomsPerGuest: number;
+	tag?: ILicenseTag;
+}
