+++ ee/app/license/definitions/ILicenseTag.ts
@@ -1,0 +1,4 @@
+export interface ILicenseTag {
+	name: string;
+	color: string;
+}
