+++ ee/server/api/licenses.ts
@@ -1,9 +1,10 @@
 import { check } from 'meteor/check';
 
-import { ILicense, getLicenses, validateFormat, flatModules, getMaxActiveUsers } from '../../app/license/server/license';
+import { getLicenses, validateFormat, flatModules, getMaxActiveUsers } from '../../app/license/server/license';
 import { Settings, Users } from '../../../app/models/server';
 import { API } from '../../../app/api/server/api';
 import { hasPermission } from '../../../app/authorization/server';
+import { ILicense } from '../../app/license/definitions/ILicense';
 
 function licenseTransform(license: ILicense): ILicense {
 	return {
