+++ ee/server/services/package.json
@@ -19,6 +19,7 @@
 	"license": "MIT",
 	"dependencies": {
 		"@rocket.chat/string-helpers": "^0.29.0",
+		"ajv": "^8.7.1",
 		"bcrypt": "^5.0.1",
 		"body-parser": "^1.19.0",
 		"colorette": "^1.3.0",
